Bloom & Updated Tonemapping
Got the bloom & HDR pass working. Also got automatic exposure working, so that for a bright scene, it darkens the image, and visa versa.

Things not quite there: auto exposure should be based on a running average to mimic eye adaptation, and there is an issue with using subpixel amounts for the bloom gaussian blur.