Decals now work for all rotations.
Decals now work in any direction, as it gets the surface tangent basis from the partial derivatives of the (relative) world position. From getting the normal of this basis, we can now do a better job of skipping pixels with a different normal value (As it is of the surface, not normal map). Furthermore, the world position relative to decal center was used to prevent new precision issues that arose from the partial derivatives.